styles: Fade in arrows
Adds a css animation to fade in the arrows on page load. Ideally, I
would like this to be slower and with a delay before beginning, however
the hover transition interrupts this, so it is necessary to pick a
suitably short animation time that it won't likely be interrupted by the
mouse.