webcam.php: Record IP address in uploaded filename
This transforms the format of uploaded images from:

  webcam-[date]-[random].jpg

to:

  [ip].[date].webcam.jpg

This effectively namespaces each separate user so prevents the need to
add the random 8 character string. This has the added benefit of being
able to diagnose abuse of the upload feature by tracking the number of
uploaded files on a per-user basis.