genetics: Remove random number pool
There are 14 uses of random numbers in the source code. After performing
performance profiles of the genetics algorithm with the random number
and with normal use of the Math.random() function, it became apparent
that there is no significant performance gain to be had from generating
the random numbers beforehand.

Closes #49.