genetics: Update "Get URL" location
The genetics page was moved from location "/genetics/" to
"/s/genetics/". While a 301 redirect rule has been added to forward
traffic correctly, the URL exported by the "Get URL" button was not. The
genetics script uses the fragment identifier to export configuration
parameters, so the exported URL:

http://chriscummins.cc/genetics/#40&25&0&2.4&10&120&6&70&1&0&1

would be forwarded to:

http://chriscummins.cc/s/genetics/

This patch updates the '#get-url' click handler to return the new
"/s/genetics/" location. A more robust solution would be to grab the
"window.location" and modify that, but this would break if someone were
to load the site locally or from a different location.