se: Implement point visualisation mechanism
This patch adds point visualisation to the Renderer. The
Renderer.addPoint() method can be used to add a visualisation point,
which is rendered using a Sphere geometry. Note that this is an
incredibly inefficient approach, and will choke most systems when
dealing with large (> 10^4) numbers of points. This will need to be
gutted and replaced with something more lightweight over time.