se: Change point size
This patch changes the point sizes so that lower points are larger than
higher points. This makes more intuitive sense since convergent search
algorithms will tend to minimise the number of points in the lower
regions, making them harder to spot; and maximising the number of points
in the higher regions, causing them to clump together and overlap.