Add support for dark mode.
This adds a bunch of media queries to set dark colors when the user
has a preference for them (set at the system level). This is
"automatic", in that no special handling is required for new content
to support it, with one exception:

A new "invertible" class can be added to images which will invert the
colors. This is useful for diagrams with a white background where the
colors aren't important, only the contrast is. For photos this will
look super weird, don't it on them. One thing to keep in mind is that
this means you should not reference the colors in invertible images,
as they may be flipped. E.g. don't say "the red bar shows the total
count".

Signed-off-by: format 20.02.05 <github.com/ChrisCummins/format>